requirements.txt
streamlit==1.28.1
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
folium==0.14.0
openpyxl==3.1.2
xlrd==2.0.1
google-generativeai==0.3.2
streamlit-chat==0.0.2
streamlit-extras==0.3.5
networkx==3.2.1
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.2
reportlab==4.0.7
Pillow==10.1.0
